+++ Planner/DistancePlanner.cpp
@@ -6,8 +6,8 @@
 #include "DistancePlanner.h"
 
 
-DistancePlanner::DistancePlanner(const WindowParamsDTO &screenParams, int numPoints, double epsilon,
-        int neighborhoodSize, double knnNeighborhoodRadiusFactor) :
+DistancePlanner::DistancePlanner(const WindowParamsDTO &screenParams, size_t numPoints, double epsilon,
+                                 int neighborhoodSize, double knnNeighborhoodRadiusFactor) :
         Planner(screenParams, numPoints, epsilon, neighborhoodSize, knnNeighborhoodRadiusFactor)
         {}
